+++ src/main/java/com/bombombom/devs/global/security/AppUserDetails.java
@@ -2,6 +2,7 @@ package com.bombombom.devs.global.security;
 
 import com.bombombom.devs.user.models.User;
 import java.util.List;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
 import org.springframework.security.core.GrantedAuthority;
@@ -9,6 +10,7 @@ import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 @Getter
+@Builder
 @ToString(exclude = "password")
 public class AppUserDetails implements UserDetails {
     private final Long id;
@@ -16,10 +18,12 @@ public class AppUserDetails implements UserDetails {
     private final String password;
     private final List<GrantedAuthority> authorities;
 
-    public AppUserDetails(User user) {
-        this.id = user.getId();
-        this.username = user.getUsername();
-        this.password = user.getPassword();
-        this.authorities = List.of(new SimpleGrantedAuthority(user.getRole().getName()));
+    public static AppUserDetails fromUser(User user) {
+        return AppUserDetails.builder()
+            .id(user.getId())
+            .username(user.getUsername())
+            .password(user.getPassword())
+            .authorities(List.of(new SimpleGrantedAuthority(user.getRole().name())))
+            .build();
     }
 }
