+++ src/test/java/com/bombombom/devs/config/TestUserDetailsServiceConfig.java
@@ -1,9 +1,13 @@
 package com.bombombom.devs.config;
 
+import com.bombombom.devs.global.security.AppUserDetails;
+import com.bombombom.devs.user.models.Role;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.context.annotation.Bean;
 import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -15,19 +19,17 @@ public class TestUserDetailsServiceConfig {
     @Bean
     public UserDetailsService testUserDetailsService() {
         return new UserDetailsService() {
-            private final List<UserDetails> users = new ArrayList<>();
+            private final List<AppUserDetails> users = new ArrayList<>();
 
             {
-                // 사용자 목록 초기화
-                users.add(User.withUsername("testuser")
-                    .password("{noop}password1")
-                    .roles("USER")
-                    .build());
-
-                users.add(User.withUsername("admin")
-                    .password("{noop}password2")
-                    .roles("ADMIN")
-                    .build());
+                users.add(
+                    AppUserDetails.builder()
+                        .id(1L)
+                        .username("testuser")
+                        .password("password")
+                        .authorities(List.of(new SimpleGrantedAuthority(Role.USER.getName())))
+                        .build()
+                );
             }
 
             @Override
