+++ src/test/java/com/bombombom/devs/user/service/UserServiceTest.java
@@ -1,17 +1,23 @@
 package com.bombombom.devs.user.service;
 
 import com.bombombom.devs.user.exception.ExistUsernameException;
+import com.bombombom.devs.user.models.Role;
+import com.bombombom.devs.user.models.User;
 import com.bombombom.devs.user.repository.UserRepository;
 import com.bombombom.devs.user.service.dto.SignupCommand;
+import com.bombombom.devs.user.service.dto.UserProfileResult;
+import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.springframework.boot.test.context.SpringBootTest;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.doReturn;
+import static org.mockito.Mockito.when;
 
 @SpringBootTest
 public class UserServiceTest {
@@ -42,5 +48,39 @@ public class UserServiceTest {
         When & Then
          */
         assertThrows(ExistUsernameException.class, () -> userService.addUser(signupCommand));
+    }
+
+    @DisplayName("username을 통해 유저를 조회할 수 있다.")
+    @Test
+    void can_retrieve_user_through_username() {
+        /*
+        Given
+         */
+        User user = User.builder()
+            .username("username")
+            .password("password")
+            .role(Role.USER)
+            .introduce("introduce")
+            .image("image")
+            .reliability(0)
+            .money(0)
+            .build();
+        when(userRepository.findById(any())).thenReturn(Optional.of(user));
+
+        /*
+        When
+         */
+        UserProfileResult result = userService.findById(user.getId());
+
+        /*
+        Then
+         */
+        assertThat(result.id()).isEqualTo(user.getId());
+        assertThat(result.username()).isEqualTo(user.getUsername());
+        assertThat(result.role()).isEqualTo(user.getRole());
+        assertThat(result.introduce()).isEqualTo(user.getIntroduce());
+        assertThat(result.image()).isEqualTo(user.getImage());
+        assertThat(result.reliability()).isEqualTo(user.getReliability());
+        assertThat(result.money()).isEqualTo(user.getMoney());
     }
 }
